autoconfig: update network device name recognission
The previous code assumed the old style network names such as eth* were
the names of interfaces. With the new udev, this is not always the case,
so this updates autoconfig to ignore lo and sit0 and assume that the
rest are valid network interfaces. Also we now use /sys/class/net/* to
detect interfaces instead of /proc/dev/net.

Reported-by: alexeyk13@yandex.ru
X-Gentoo-Bug: 471054
X-Gentoo-Bug-URL: https://bugs.gentoo.org/show_bug.cgi?id=471054