net-setup: Fix zero argument mode
The grepped output of 'ifconfig -a' included a colon after each interface.
This broke subsequent calls for information on the interface.

Also, instead of flashing an error when a path did not exist,
check that it does exist before calling basename.

Skip sit0 as it cannot be configured.

Finally, check the return value of the dialog call.
Only zero means an interface was selected to be processed.